refactor: move config out of type-extensions
Swap to a single `./type-extensions.ts` file. We only add config in the
plugin.
